--- src/main/java/com/ll/codicaster/boundedContext/image/service/ImageService.java
@@ -1,8 +1,0 @@
-package com.ll.codicaster.boundedContext.image.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class ImageService {
-
-}
