+++ src/main/resources/application.yml
@@ -6,7 +6,7 @@ spring:
     driver-class-name: org.mariadb.jdbc.Driver
     url: jdbc:mariadb://127.0.0.1:3306/codicaster_dev?useUnicode=true&characterEncoding=utf8&autoReconnect=true&serverTimezone=Asia/Seoul
     username: root
-    password:
+    password: 1234
   security:
     oauth2:
       client:
@@ -57,4 +57,4 @@ custom:
   site:
     baseUrl: http://localhost:8080
 file:
-  upload-dir: C:\Users\82102\IdeaProjects\CodiCaster__-main\images
+  upload-dir: /Users/hojun/Desktop/CodiCaster/images
